+++ src/middleware.ts
@@ -1,5 +1,5 @@
 import { type NextRequest } from "next/server";
-import { updateSession } from "@/utils/supabase/middleware";
+import { updateSession } from "@/shared/utils/supabase/middleware";
 
 export async function middleware(request: NextRequest) {
   return await updateSession(request);
