+++ src/pages/api/auth/confirm.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next/";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/shared/utils/supabase/server";
 
 export default async function handler(
   req: NextApiRequest,
