+++ src/pages/auth/register/index.tsx
@@ -3,9 +3,9 @@ import { useForm, Controller, type SubmitHandler } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Box, Typography, TextField, Button, Paper, Link } from "@mui/material";
 import NextLink from "next/link";
-import { signUpSchema, type signUpValues } from "@/schemas/auth";
+import { signUpSchema, type signUpValues } from "@/shared/schemas/auth";
 import { ConfirmEmailModal } from "../../../components/ConfirmEmailModal";
-import { createClient } from "@/utils/supabase/client";
+import { createClient } from "@/shared/utils/supabase/client";
 
 export default function SignUpForm() {
   const [generalError, setGeneralError] = useState("");
