+++ src/shared/types/database.type.ts
