+++ src/shared/types/postgrest-types.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/utils/supabase/client";
+import { createClient } from "@/shared/utils/supabase/client";
 
 const supabase = createClient();
