+++ src/shared/utils/supabase/server.ts
@@ -2,7 +2,7 @@
 import { createServerClient, type CookieOptions } from "@supabase/ssr";
 import type { NextApiRequest, NextApiResponse } from "next"; // Import NextApiRequest and NextApiResponse
 import { serialize } from "cookie"; // Import serialize and parse from 'cookie'
-import type { Database } from "@/types/database.type";
+import type { Database } from "@/shared/types/database.type";
 
 // This createClient function is now specifically for Pages Router API routes
 export const createClient = (req: NextApiRequest, res: NextApiResponse) => {
